change data types from Doubles to BigDecimals

+++ src/main/java/pl/kopacz/repository/SupplyRepository.java
@@ -4,11 +4,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import pl.kopacz.domain.Spice;
 import pl.kopacz.domain.Supply;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @SuppressWarnings("unused")
 public interface SupplyRepository extends JpaRepository<Supply, Long> {
 
-    List<Supply> findBySpiceAndAmountGreaterThanOrderByIdAsc(Spice spice, Double amount);
+    List<Supply> findBySpiceAndAmountGreaterThanOrderByIdAsc(Spice spice, BigDecimal amount);
 
 }
